Introduce an iterator over all inlined functions
The inlined function index is exposed to users of libdrgn by means
of the newly introduced `drgn_inlined_functions_iterator`. The decision
to provide this as an iterator rather than a mapping was motivated by
the fact that the mapping which `drgn` uses internally is not likely to
be very useful to consumers, seeing as its keyed by DWARF DIE addresses,
and building a secondary mapping keyed on something more user-friendly
(such as function name/linkage-name) would be extremely expensive, and
is not something we should force consumers to pay for in terms of
runtime performance if they do not need this functionality.

Signed-off-by: Kevin Svetlitski <svetlitski@fb.com>