Prevent `drgn_object_from_dwarf` from failing unnecessarily
Previously the function would fail if it could not initialize a
`drgn_object_locator` for the given DIE, but I didn't realize that
`drgn_object_from_dwarf` was used in other contexts where you don't
actually care about locating the object. As such, in those contexts
an error in initializing the `drgn_object_locator` is now ignored.