Conditionally disable problematic behavior w.r.t resolving incomplete types in classes
Indexing classes as if they were namespaces for the sake of resolving
incomplete types within them would cause non-deterministic behavior as a
consequence of introducing concurrent modification (the same DIE gets
put into multiple shards, resulting in concurrent modification even
though each shard is guarded by a lock). For the sake of getting the
type iterator API working as quickly as possible, we disable this
problematic behavior through an environment variable. Note that this
means that when the environment variable is set, incomplete types within
classes will NOT be resolved.