Return multiple matching functions from `drgn_oil_type_iterator_next()`
This lets us compile OIL with `-O0`.

There are overloads of `CodegenHandler::getObjectSize()` with different
arguments. When compiling with optimizations enabled, unused versions
will not be present in the binary, so we would only match the overload
being used. However, with optimizations turned off, all overloads will
be present in the binary.

Change the interface of `drgn_oil_type_iterator_next()` so that it can
return all the present overloads of `getObjectSize()`.