libdrgn: proof of concept DWO support
The general approach is: in drgn_dwarf_index_read_cus(), for each CU,
ask libdw for the "sub-DIE". For skeleton CUs, this is the split CU DIE
from the .dwo file. From that Dwarf_Die, we can get the Dwarf_CU and
then the Dwarf handle. Then, we wrap that in a struct drgn_elf_file
(cached in a hash table in the struct drgn_module), which the DWARF
index can work with from there.

The biggest TODO (other than cleaning up the code) is to handle the file
name table; we need to get that from the skeleton CU, not the split CU
that we're indexing. There are likely other places outside of the DWARF
index that need to be fixed to get the information from the main debug
file for stuff that stays there in Debug Fission.